+++ src/main/java/de/raidcraft/items/crafting/recipes/CustomShapedRecipe.java
@@ -100,6 +100,7 @@ public class CustomShapedRecipe extends ShapedRecipe implements CustomRecipe {
             // create new
             recipe = new TCraftingRecipe();
         }
+        recipe.setName(getName());
         recipe.setResult(RaidCraft.getItemIdString(getResult()));
         recipe.setAmount(getResult().getAmount());
         recipe.setShape(StringUtils.join(getShape(), "|"));
@@ -107,7 +108,9 @@ public class CustomShapedRecipe extends ShapedRecipe implements CustomRecipe {
         recipe.setPermission(getPermission());
         database.save(recipe);
         // create the ingredients
-        database.delete(recipe.getIngredients());
+        if (recipe.getIngredients() != null && !recipe.getIngredients().isEmpty()) {
+            database.delete(recipe.getIngredients());
+        }
         for (Map.Entry<Character, ItemStack> entry : getIngredientMap().entrySet()) {
             TCraftingRecipeIngredient ingredient = new TCraftingRecipeIngredient();
             ingredient.setRecipe(recipe);
